+++ nw/ARENA/TimeRuntimeState.cpp
@@ -3,5 +3,6 @@
 #include "h/super.h"
 #include "storage/h/savefile.h"
 #include "suavik.h"
+#include "TimeRuntimeState.h"
 
 #include "TimeRuntimeState.inl"